+++ sdk/search/Azure.Search.Documents/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Removed constructor from `SynonymMap` with `IEnumerable<string>` parameter.
 - `SearchServiceClient.GetIndexes` and `SearchServiceClient.GetIndexesAsync` now return `Pageable<SearchIndex>` and `AsyncPageable<SearchIndex>` respectively.
+- Combined `MatchConditions` and `SearchRequestOptions` into `SearchConditionalOptions`.
 
 ## 1.0.0-preview.2 (2020-04-06)